OBPIH-2831 Minor improvements to Ship Order page

+++ grails-app/controllers/org/pih/warehouse/order/OrderController.groovy
@@ -94,7 +94,7 @@ class OrderController {
                                     orderItem: orderItem,
                                     shipmentItem: shipmentItem,
                                     quantityMinimum: 0,
-                                    quantityToShip: shipmentItem.quantity / orderItem?.quantityPerUom,
+                                    quantityToShip: (shipmentItem.quantity / orderItem?.quantityPerUom) as int,
                                     quantityMaximum: orderItem.quantityRemaining
 
                             )
